AbstractRemoveFieldsFromObjectTypeHookSet
