+++ layers/Engine/packages/component-model/src/Hooks/AbstractRemoveIDAndSelfFieldsFromObjectTypeHookSet.php
@@ -6,11 +6,8 @@ namespace PoP\ComponentModel\Hooks;
 
 use PoP\ComponentModel\FieldResolvers\InterfaceType\InterfaceTypeFieldResolverInterface;
 use PoP\ComponentModel\FieldResolvers\ObjectType\ObjectTypeFieldResolverInterface;
-use PoP\ComponentModel\TypeResolvers\HookHelpers;
 use PoP\ComponentModel\TypeResolvers\InterfaceType\InterfaceTypeResolverInterface;
 use PoP\ComponentModel\TypeResolvers\ObjectType\ObjectTypeResolverInterface;
-use PoP\Root\App;
-use PoP\Root\Hooks\AbstractHookSet;
 
 /**
  * To be used together with:
@@ -18,46 +15,18 @@ use PoP\Root\Hooks\AbstractHookSet;
  * - RemoveNodeInterfaceObjectTypeResolverTrait
  * - AbstractTransientObject
  */
-abstract class AbstractRemoveIDAndSelfFieldsFromObjectTypeHookSet extends AbstractHookSet
+abstract class AbstractRemoveIDAndSelfFieldsFromObjectTypeHookSet extends AbstractRemoveFieldsFromObjectTypeHookSet
 {
-    protected function init(): void
-    {
-        App::addFilter(
-            HookHelpers::getHookNameToFilterField(),
-            $this->filterFields(...),
-            10,
-            4
-        );
-    }
-
-    public function filterFields(
-        bool $include,
+    protected function matchesCondition(
         ObjectTypeResolverInterface | InterfaceTypeResolverInterface $objectTypeOrInterfaceTypeResolver,
         ObjectTypeFieldResolverInterface | InterfaceTypeFieldResolverInterface $objectTypeOrInterfaceTypeFieldResolver,
         string $fieldName
     ): bool {
-        if (
-            ($fieldName !== 'id' && $fieldName !== 'self')
-            || ($fieldName === 'self' && !$this->removeSelfField())
-        ) {
-            return $include;
-        }
-
-        $objectTypeOrInterfaceTypeResolverClass = $this->getObjectTypeOrInterfaceTypeResolverClass();
-        if (is_a($objectTypeOrInterfaceTypeResolver, $objectTypeOrInterfaceTypeResolverClass, true)) {
-            return false;
-        }
-
-        return $include;
+        return $fieldName === 'id' || ($fieldName === 'self' && $this->removeSelfField());
     }
 
     protected function removeSelfField(): bool
     {
         return true;
     }
-
-    /**
-     * @phpstan-return class-string<ObjectTypeResolverInterface|InterfaceTypeResolverInterface>
-     */
-    abstract protected function getObjectTypeOrInterfaceTypeResolverClass(): string;
 }
